refactor: rearrange project files
