+++ src/aepp-wallet-communication/connection-proxy.ts
