+++ src/aepp-wallet-communication/connection/BrowserWindowMessage.ts
@@ -17,7 +17,7 @@
 
 import BrowserConnection from './Browser'
 import { MESSAGE_DIRECTION } from '../schema'
-import { InternalError, RpcConnectionError } from '../../errors'
+import { InternalError, RpcConnectionError } from '../../utils/errors'
 
 export type ImplPostMessage = Pick<Window, 'addEventListener' | 'removeEventListener' | 'postMessage'>
