+++ src/index.ts
@@ -25,27 +25,27 @@ export * from './tx/builder/helpers'
 export * from './tx/builder/schema'
 export * from './utils/amount-formatter'
 export * from './utils/hd-wallet'
-export * from './ae/aens'
-export * from './ae/contract'
-export * from './ae/oracle'
-export * from './ae/spend'
+export * from './aens'
+export * from './contract/methods'
+export * from './oracle'
+export * from './spend'
 export * from './contract/ga'
 
 export { default as AeSdkBase } from './AeSdkBase'
 export { default as AeSdk } from './AeSdk'
 export { default as AeSdkAepp } from './AeSdkAepp'
 export { default as AeSdkWallet } from './AeSdkWallet'
-export { default as Node } from './node'
+export { default as Node } from './Node'
 export { default as verifyTransaction } from './tx/validator'
 export { default as calculateMinFee } from './tx/min-fee'
-export { default as AccountBase } from './account/base'
-export { default as MemoryAccount } from './account/memory'
-export { default as Compiler } from './contract/compiler'
+export { default as AccountBase } from './account/Base'
+export { default as MemoryAccount } from './account/Memory'
+export { default as Compiler } from './contract/Compiler'
 export { default as Channel } from './channel'
 
-export { default as connectionProxy } from './utils/aepp-wallet-communication/connection-proxy'
-export * from './utils/aepp-wallet-communication/schema'
-export { default as walletDetector } from './utils/aepp-wallet-communication/wallet-detector'
-export { default as BrowserRuntimeConnection } from './utils/aepp-wallet-communication/connection/BrowserRuntime'
-export { default as BrowserWindowMessageConnection } from './utils/aepp-wallet-communication/connection/BrowserWindowMessage'
+export { default as connectionProxy } from './aepp-wallet-communication/connection-proxy'
+export * from './aepp-wallet-communication/schema'
+export { default as walletDetector } from './aepp-wallet-communication/wallet-detector'
+export { default as BrowserRuntimeConnection } from './aepp-wallet-communication/connection/BrowserRuntime'
+export { default as BrowserWindowMessageConnection } from './aepp-wallet-communication/connection/BrowserWindowMessage'
 export * from './utils/errors'
