+++ src/spend.ts
@@ -14,15 +14,15 @@
  *  OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
  *  PERFORMANCE OF THIS SOFTWARE.
  */
-import { sendTransaction, getAccount, getBalance, resolveName } from '../chain'
-import { _buildTx, BuildTxOptions } from '../tx'
-import { buildTxHash, unpackTx } from '../tx/builder'
+import { sendTransaction, getAccount, getBalance, resolveName } from './chain'
+import { _buildTx, BuildTxOptions } from './tx'
+import { buildTxHash, unpackTx } from './tx/builder'
 import BigNumber from 'bignumber.js'
-import { ArgumentError } from '../utils/errors'
-import { EncodedData } from '../utils/encoder'
-import { createMetaTx } from '../contract/ga'
-import { TX_TYPE, AensName } from '../tx/builder/schema'
-import AccountBase from '../account/base'
+import { ArgumentError } from './utils/errors'
+import { EncodedData } from './utils/encoder'
+import { createMetaTx } from './contract/ga'
+import { TX_TYPE, AensName } from './tx/builder/schema'
+import AccountBase from './account/Base'
 
 /**
  * Sign and post a transaction to the chain
