+++ src/tx/index.ts
@@ -29,7 +29,7 @@ import {
   ArgumentError, UnsupportedProtocolError, UnknownTxError, InvalidTxParamsError
 } from '../utils/errors'
 import BigNumber from 'bignumber.js'
-import Node from '../node'
+import Node from '../Node'
 import { EncodedData } from '../utils/encoder'
 import { buildTx as syncBuildTx, unpackTx } from './builder/index'
 import calculateMinFee from './min-fee'
