+++ test/integration/ga.ts
@@ -20,7 +20,7 @@ import { expect } from 'chai'
 import { getSdk } from '.'
 import { generateKeyPair } from '../../src/utils/crypto'
 import { encode } from '../../src/utils/encoder'
-import MemoryAccount from '../../src/account/memory'
+import MemoryAccount from '../../src/account/Memory'
 import { unpackTx } from '../../src/tx/builder'
 import { ContractInstance } from '../../src/contract/aci'
 import { AeSdk, TX_TYPE, UnexpectedTsError } from '../../src'
