+++ test/integration/oracle.ts
@@ -19,10 +19,10 @@ import { describe, it, before } from 'mocha'
 import { expect } from 'chai'
 import { getSdk } from '.'
 import { generateKeyPair } from '../../src/utils/crypto'
-import MemoryAccount from '../../src/account/memory'
+import MemoryAccount from '../../src/account/Memory'
 import { ORACLE_TTL_TYPES, QUERY_FEE } from '../../src/tx/builder/schema'
 import { decode } from '../../src/tx/builder/helpers'
-import { postQueryToOracle, registerOracle } from '../../src/ae/oracle'
+import { postQueryToOracle, registerOracle } from '../../src/oracle'
 import { EncodedData } from '../../src/utils/encoder'
 import { AeSdk, UnexpectedTsError } from '../../src'
