+++ test/integration/txVerification.ts
@@ -2,7 +2,7 @@ import { before, describe, it } from 'mocha'
 import { expect } from 'chai'
 import { getSdk } from '.'
 import { generateKeyPair } from '../../src/utils/crypto'
-import MemoryAccount from '../../src/account/memory'
+import MemoryAccount from '../../src/account/Memory'
 import verifyTransaction from '../../src/tx/validator'
 import { InvalidTxError, InvalidTxParamsError } from '../../src/utils/errors'
 import { TX_TYPE } from '../../src/tx/builder/schema'
